+++ jolt-atlas-core/src/onnx_proof/ops/constant.rs
@@ -2,7 +2,7 @@ use crate::onnx_proof::{
     ops::{OperatorProofTrait, Prover, Verifier},
     ProofId,
 };
-use crate::utils::opening_id_builder::AccOpeningAccessor;
+use crate::utils::opening_access::AccOpeningAccessor;
 use atlas_onnx_tracer::{node::ComputationNode, ops::Constant};
 
 use joltworks::{
