+++ jolt-atlas-core/src/onnx_proof/ops/rsqrt.rs
@@ -7,7 +7,7 @@ use crate::{
         },
         ProofId, ProofType, Prover, Verifier,
     },
-    utils::opening_id_builder::{AccOpeningAccessor, Target},
+    utils::opening_access::{AccOpeningAccessor, Target},
 };
 use atlas_onnx_tracer::{
     model::trace::{LayerData, Trace},
