+++ jolt-atlas-core/src/onnx_proof/ops/softmax_last_axis/exp_sum.rs
@@ -1,4 +1,4 @@
-use crate::utils::opening_id_builder::AccOpeningAccessor;
+use crate::utils::opening_access::AccOpeningAccessor;
 use atlas_onnx_tracer::node::ComputationNode;
 use common::parallel::par_enabled;
 use common::VirtualPoly;
