+++ jolt-atlas-core/src/onnx_proof/ops/softmax_last_axis/recip_mult.rs
@@ -1,4 +1,4 @@
-use crate::utils::opening_id_builder::AccOpeningAccessor;
+use crate::utils::opening_access::AccOpeningAccessor;
 use atlas_onnx_tracer::node::ComputationNode;
 use common::VirtualPoly;
 use joltworks::{
