+++ jolt-atlas-core/src/utils/mod.rs
@@ -7,7 +7,7 @@ use joltworks::utils::{interleave_bits, lookup_bits::LookupBits};
 use rayon::prelude::*;
 
 pub mod dims;
-pub mod opening_id_builder;
+pub mod opening_access;
 
 /// Computes the floored (Euclidean-style) remainder of `a / b`.
 ///
